feat: finished flow fundamentals
